Move python finding logic out of catalogue and into manager where it belongs
